[FD-181] feat: showToast 수정하여 내용만 입력해도 띄워지도록 수정
